upgrade inference to multi
